backend/requirements.txt
fastapi
uvicorn
pandas
pydantic
python-multipart
python-dotenv
requests